[WIP] lower_snake_case naming + CMakeLists
