--- vision-be/vision-panettone/vision/RepositoryFactory/RepositoryFactoryMongoDB/RepositoryFactoryMongoDB.cpp
@@ -1,9 +1,0 @@
-// RepositoryFactoryMongoDB.cpp
-#include "RepositoryFactoryMongoDB.h"
-
-#include "../../data/Frame/FrameRepositoryMongoDB/FrameRepositoryMongoDB.h"
-#include "../../data/Frame/IFrameRepository.h"
-
-std::unique_ptr<IFrameRepository> RepositoryFactoryMongoDB::createFrameRepository() const {
-  return std::make_unique<FrameRepositoryMongoDB>();
-}
