+++ vision-be/vision-panettone/vision/data/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory(frame)
