+++ vision-be/vision-panettone/vision/data/frame/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_subdirectory(frame_repository_mongo_db)
+
+robocin_cpp_library(
+  NAME frame_repository
+  DEPS i_frame_repository.h ../../src/frame/frame.h
+)
