+++ vision-be/vision-panettone/vision/data/frame/frame_repository_mongo_db/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME frame_repository_mongo_db
+  SRCS frame_repository_mongo_db.cpp
+  DEPS frame_repository_mongo_db.h ../../../src/frame/frame.h ../i_frame_repository.h
+)
