+++ vision-be/vision-panettone/vision/data/frame/frame_repository_mongo_db/frame_repository_mongo_db.h
@@ -1,0 +1,26 @@
+// FrameRepositoryMongoDB.h
+#ifndef FRAME_REPOSITORY_MONGODB_H
+#define FRAME_REPOSITORY_MONGODB_H
+
+#include "../../../src/frame/frame.h"
+#include "../i_frame_repository.h"
+
+#include <mongocxx/client.hpp>
+
+class FrameRepositoryMongoDB : public IFrameRepository {
+ public:
+  FrameRepositoryMongoDB();
+
+  void save(const Frame& frame) override;
+  [[nodiscard]] std::optional<Frame> find(const std::string& frame_id) override;
+  std::optional<Frame> update(const std::string& frame_id,
+                              bsoncxx::document::value& update) override;
+  void remove(const std::string& frame_id) override;
+
+ private:
+  mongocxx::client client_;
+  mongocxx::database db_;
+  mongocxx::collection collection_;
+};
+
+#endif // FRAME_REPOSITORY_MONGODB_H
