+++ vision-be/vision-panettone/vision/repository_factory/repository_factory_mapping/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME repository_factory_mapping
+  SRCS repository_factory_mapping.cpp
+  DEPS repository_factory_mapping.h
+)
