+++ vision-be/vision-panettone/vision/repository_factory/repository_factory_mongo_db/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME repository_factory_mongo_db
+  SRCS repository_factory_mongo_db.cpp
+  DEPS repository_factory_mongo_db.h
+)
