--- vision-be/vision-panettone/vision/src/Robot/Color/Color.cpp
@@ -1,13 +1,0 @@
-// Color.cpp
-#include "Color.h"
-
-#include <ostream>
-
-std::ostream& operator<<(std::ostream& os, const Color& color) {
-  switch (color) {
-    case Color::BLUE: os << "BLUE"; break;
-    case Color::YELLOW: os << "YELLOW"; break;
-    default: os << "Unknown Color";
-  }
-  return os;
-}
