--- vision-be/vision-panettone/vision/src/Source/Source.cpp
@@ -1,11 +1,0 @@
-// Source.cpp
-#include "Source.h"
-
-std::ostream& operator<<(std::ostream& os, const Source& source) {
-  switch (source) {
-    case Source::SIMULATION: os << "SIMULATION"; break;
-    case Source::SSLVISION: os << "SSLVISION"; break;
-    default: os << "Unknown Source";
-  }
-  return os;
-}
