+++ vision-be/vision-panettone/vision/src/ball/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME ball
+  SRCS ball.cpp
+  DEPS ball.h
+)
