+++ vision-be/vision-panettone/vision/src/field/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME field
+  SRCS field.cpp
+  DEPS field.h
+)
