+++ vision-be/vision-panettone/vision/src/frame/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME frame
+  SRCS frame.cpp
+  DEPS frame.h
+)
