+++ vision-be/vision-panettone/vision/src/point/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory(point_2df)
+
+add_subdirectory(point_3df)
