+++ vision-be/vision-panettone/vision/src/point/point_2df/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME point_2df
+  SRCS point_2df.cpp
+  DEPS point_2df.h
+)
