+++ vision-be/vision-panettone/vision/src/point/point_3df/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME point_3df
+  SRCS point_3df.cpp
+  DEPS point_3df.h
+)
