+++ vision-be/vision-panettone/vision/src/robot/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_subdirectory(color)
+
+robocin_cpp_library(
+  NAME robot
+  SRCS robot.cpp
+  DEPS robot.h
+)
