+++ vision-be/vision-panettone/vision/src/robot/color/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME robot_color
+  SRCS color.cpp
+  DEPS color.h
+)
