+++ vision-be/vision-panettone/vision/src/robot/color/color.h
