+++ vision-be/vision-panettone/vision/src/source/CMakeLists.txt
@@ -1,0 +1,5 @@
+robocin_cpp_library(
+  NAME source
+  SRCS source.cpp
+  DEPS source.h
+)
