+++ vision-be/vision-panettone/vision/src/source/source.cpp
@@ -1,0 +1,11 @@
+// Source.cpp
+#include "source.h"
+
+std::ostream& operator<<(std::ostream& os, const Source& source) {
+  switch (source) {
+    case Source::SIMULATION: os << "SIMULATION"; break;
+    case Source::SSLVISION: os << "SSLVISION"; break;
+    default: os << "Unknown Source";
+  }
+  return os;
+}
